feat: 주문표 Receipt 컴포넌트화 완료
