--- frontend/src/components/DetailMenuBar/atom.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-const menuTab = atom({
-  key: 'menuTab',
-  default: '메뉴',
-});
-
-export { menuTab };
