+++ frontend/src/components/DetailMenuBar/styles.js
@@ -27,120 +27,4 @@ const Section = styled.div`
   }
 `;
 
-const Menu = styled.div`
-  display: flex;
-  justify-content: space-between;
-  background-color: white;
-  height: 104px;
-  padding: 10px 10px 10px;
-  border-bottom: 1px solid grey;
-`;
-
-const MenuInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-`;
-
-const MenuImg = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 10px;
-  > img {
-    width: 80px;
-    height: 80px;
-  }
-`;
-
-const Rating = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 40px;
-  background-color: ${(e) => e.theme.mainColor};
-  height: 100px;
-  width: 100%;
-`;
-
-const Comment = styled.div`
-  color: black;
-  background-color: lightgrey;
-  font-size: 14px;
-  padding: 5px;
-  line-height: 20px;
-  padding-bottom: 20px;
-  border-bottom: 1px solid grey;
-`;
-
-const NameDateReply = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const ReplyDeleteBtn = styled.button`
-  background-color: lightgrey;
-  margin-left: 20px;
-  font-size: 16px;
-  font-weight: 700;
-  border: 0px;
-`;
-
-const ThumnailBox = styled.button`
-  flex-direction: row;
-`;
-
-const Thumnail = styled.img`
-  width: 200px;
-  height: 120px;
-`;
-
-const TruckInfoBody = styled.div`
-  margin: 20px;
-  line-height: 30px;
-`;
-
-const TruckInfoTitle = styled.div`
-  font-size: 25px;
-  font-weight: 900;
-  color: black;
-  border-bottom: 2px solid black;
-  padding-bottom: 10px;
-`;
-
-const TruckInfoContent = styled.div`
-  white-space: pre-wrap;
-  color: black;
-  padding-bottom: 50px;
-  padding-top: 10px;
-`;
-
-const TruckInfoContentKey = styled.div`
-  color: grey;
-  padding-bottom: 15px;
-`;
-
-const TruckInfoContentValue = styled.span`
-  color: black;
-`;
-
-export {
-  MenuBar,
-  TabBtn,
-  Section,
-  Menu,
-  MenuInfo,
-  MenuImg,
-  Rating,
-  Comment,
-  NameDateReply,
-  ReplyDeleteBtn,
-  ThumnailBox,
-  Thumnail,
-  TruckInfoBody,
-  TruckInfoTitle,
-  TruckInfoContent,
-  TruckInfoContentKey,
-  TruckInfoContentValue,
-};
+export { MenuBar, TabBtn, Section };
