+++ frontend/src/pages/FoodList/FoodDetail.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import DetailMenuBar from '../../components/DetailMenuBar';
+import { DetailMenuBar, Receipt } from '../../components';
 import {
   Section,
   MainBody,
@@ -8,15 +8,7 @@ import {
   FoodTruckImg,
   CapsulizedInfo,
   Notice,
-  StickyBody,
-  CartListBody,
-  Cart,
-  CartTab,
-  CartList,
-  CartListAdd,
-  TotalPrice,
-  OrderBtn,
-} from './FoodDetailstyles';
+} from './styles';
 
 function FoodDetail() {
   return (
@@ -44,58 +36,7 @@ function FoodDetail() {
 
         <DetailMenuBar />
       </MainBody>
-      <StickyBody>
-        <Cart>
-          <CartTab>
-            <div>장바구니</div>
-            <button type="button">리셋</button>
-          </CartTab>
-          <CartListBody>
-            <CartList>
-              <div>햄버거</div>
-
-              <CartListAdd>
-                <div>8,000원</div>
-                <div className="수량버튼">
-                  <button type="button">-</button>
-                  <span>1</span>
-                  <button type="button">+</button>
-                </div>
-              </CartListAdd>
-            </CartList>
-            <CartList>
-              <div>피자</div>
-
-              <CartListAdd>
-                <div>18,000원</div>
-                <div className="수량버튼">
-                  <button type="button">-</button>
-                  <span>1</span>
-                  <button type="button">+</button>
-                </div>
-              </CartListAdd>
-            </CartList>
-            <CartList>
-              <div>치킨</div>
-
-              <CartListAdd>
-                <div>12,000원</div>
-                <div className="수량버튼">
-                  <button type="button">-</button>
-                  <span>1</span>
-                  <button type="button">+</button>
-                </div>
-              </CartListAdd>
-            </CartList>
-          </CartListBody>
-          <TotalPrice className="totalPrice">합계: 30,000원</TotalPrice>
-        </Cart>
-
-        <OrderBtn>
-          <button type="button">주문 취소</button>
-          <button type="button">바로 주문하기</button>
-        </OrderBtn>
-      </StickyBody>
+      <Receipt />
     </Section>
   );
 }
